Rename QueryBuilder component to DataExplorer
